+++ src/Dto/AuthorizationFailReason.php
@@ -1,6 +1,6 @@
 <?php
 //
-// Copyright (C) 2018 Authlete, Inc.
+// Copyright (C) 2018-2020 Authlete, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -186,6 +186,21 @@ class AuthorizationFailReason
      * @var AuthorizationFailReason
      */
     public static $INTERACTION_REQUIRED;
+
+
+    /**
+     * The requested resource is invalid, missing, unknown, or malformed.
+     *
+     * Using this reason will result in `error=invalid_target`.
+     *
+     * @static
+     * @var AuthorizationFailReason
+     *
+     * @see https://www.rfc-editor.org/rfc/rfc8707.html RFC 8707 Resource Indicators for OAuth 2.0
+     *
+     * @since 1.8
+     */
+    public static $INVALID_TARGET;
 }